Replace email with username in login and registration
Updated login and registration forms to use a username field instead of an email field. Removed the `Email` property from the login model and added a `Username` property. Adjusted authentication logic to utilize the username for login. Similar changes made in the registration model to accommodate the new username field.